refactor: renamed useGetMainMarketHistory to useGetMarketHistory
