+++ src/clients/api/index.ts
@@ -211,9 +211,9 @@ export { default as getIsolatedPools } from './queries/getIsolatedPools';
 export * from './queries/getIsolatedPools';
 export { default as useGetIsolatedPools } from './queries/getIsolatedPools/useGetIsolatedPools';
 
-export { default as getMainMarketHistory } from './queries/getMainMarketHistory';
-export * from './queries/getMainMarketHistory';
-export { default as useGetMainMarketHistory } from './queries/getMainMarketHistory/useGetMainMarketHistory';
+export { default as getMarketHistory } from './queries/getMarketHistory';
+export * from './queries/getMarketHistory';
+export { default as useGetMarketHistory } from './queries/getMarketHistory/useGetMarketHistory';
 
 export { default as getVTokenInterestRateModel } from './queries/getVTokenInterestRateModel';
 export * from './queries/getVTokenInterestRateModel';
